requirements.txt
requests
openai
python-docx
python-dotenv
httpx
